+++ CHANGELOG.md
@@ -5,8 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- No changes yet.
-
+### Added
+- added `AddHeader` as a shorthand to add a single response header.
 
 ## [1.60.0] - 2022-02-03
 ### Added